+++ R/riverspace.R
@@ -18,6 +18,9 @@
 #' @examplesIf interactive()
 #' bucharest_osm <- get_osm_example_data()
 #' delineate_riverspace(bucharest_osm$river_surface, bucharest_osm$buildings)
+#' @srrstats {G2.6} One-dimensional distance input is pre-processed by
+#'   `preprocess_distance()` to handle `units` objects or other vector-like
+#'   classes with storage mode `numeric`.
 #' @srrstats {G2.7} The `river` and `occluders` parameters accept
 #'   domain-specific tabular input of type `sf`.
 #' @srrstats {SP4.0, SP4.0b, SP4.1, SP4.2} The return value is of class
@@ -25,6 +28,8 @@
 #'   same units as the input.
 delineate_riverspace <- function(river, occluders = NULL, density = 1 / 50,
                                  ray_num = 40, ray_length = 100) {
+  # Pre-process distance input
+  ray_length <- preprocess_distance(ray_length)
   # Check input
   checkmate::assert_multi_class(river, c("list", "sf", "sfc"))
   checkmate::assert_vector(river, min.len = 1)