+++ R/srr-stats-standards.R
@@ -67,10 +67,6 @@ NULL
 #' @srrstatsNA {G1.6} As there are no alternative implementations, no
 #'   performance claims are made in this package.
 #' @srrstatsNA {G2.4d, G2.4e, G2.5} This package does not make use of factors.
-#' @srrstatsNA {G2.6} This package only accepts one-dimensional inputs
-#'   inheriting from base vector classes. Input types are strictly validated
-#'   with `checkmate`, and custom vector-like classes are not in the list of
-#'   accepted input classes. Therefore, this standard does not apply.
 #' @srrstatsNA {G2.11, G2.12} This package does not utilize list columns or
 #'   columns with non-standard class attributes in `data.frame`-like objects.
 #' @srrstatsNA {G2.14, G2.14a, G2.14b, G2.14c} These standards are not